Renamed helpers to setup, added parseJSONResponse & TestMain
