+++ internal/tests/helpers.go
@@ -1,122 +1,89 @@
 package tests
 
 import (
-	"context"
+	"bytes"
+	"encoding/json"
 	"fmt"
-	"log/slog"
+	"io"
+	"net/http"
+	"net/url"
 	"testing"
-
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/testcontainers/testcontainers-go/modules/compose"
 )
 
-type TestResources struct {
-	ComposeStack *compose.DockerCompose
-	KeyDbPool    *pgxpool.Pool
-	AppUrl       string
-}
-
-// setupTestResources creates and starts all required containers for testing
-func setupTestResources(ctx context.Context, t *testing.T) (*TestResources, error) {
+func createShortUrl(t *testing.T, baseUrl, longUrl string) *http.Response {
 	t.Helper()
-	composeStack, err := compose.NewDockerCompose("./docker-compose.yaml")
+
+	urlString := fmt.Sprintf("%s/create", baseUrl)
+	parsedUrl, err := url.Parse(urlString)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create compose stack: %w", err)
+		t.Fatalf("failed to parse POST /create url: %v", err)
 	}
 
-	err = composeStack.Up(ctx, compose.Wait(true))
-	defer func() {
-		// handle cleanup here if setup fails halfway through
-		if err != nil {
-			cleanupErr := composeStack.Down(ctx, compose.RemoveOrphans(true), compose.RemoveImagesLocal)
-			slog.Error("cleanup error", slog.Any("error", cleanupErr))
-		}
-	}()
+	b, err := json.Marshal(map[string]string{
+		"url": longUrl,
+	})
 	if err != nil {
-		return nil, fmt.Errorf("failed to start compose stack: %w", err)
+		t.Fatalf("failed to parse POST /create request body: %v", err)
 	}
 
-	appUrl, err := getAppUrl(ctx, composeStack)
+	req, err := http.NewRequestWithContext(t.Context(), http.MethodPost, parsedUrl.String(), bytes.NewBuffer(b))
 	if err != nil {
-		return nil, err
+		t.Fatalf("failed to create POST /create request: %v", err)
 	}
 
-	keyDbUrl, err := getKeyDbUrl(ctx, composeStack)
+	req.Header.Set("Content-Type", "application/json")
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
 	if err != nil {
-		return nil, err
+		t.Fatalf("failed to make POST /create request: %v", err)
 	}
 
-	keyDbPool, err := pgxpool.New(ctx, keyDbUrl)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create new key db pool: %w", err)
-	}
-
-	if err := keyDbPool.Ping(ctx); err != nil {
-		return nil, fmt.Errorf("failed to connect to key db: %w", err)
-	}
-
-	return &TestResources{
-		ComposeStack: composeStack,
-		AppUrl:       appUrl,
-		KeyDbPool:    keyDbPool,
-	}, nil
+	return resp
 }
 
-func (tr *TestResources) Cleanup(ctx context.Context, t *testing.T) {
-	if tr == nil {
-		return
+func getLongUrl(t *testing.T, baseUrl, key string) *http.Response {
+	t.Helper()
+
+	urlString := fmt.Sprintf("%s/%s", baseUrl, key)
+	parsedUrl, err := url.Parse(urlString)
+	if err != nil {
+		t.Fatalf("failed to parse GET /%s url: %v", key, err)
 	}
 
-	if tr.KeyDbPool != nil {
-		tr.KeyDbPool.Close()
+	req, err := http.NewRequestWithContext(t.Context(), http.MethodGet, parsedUrl.String(), http.NoBody)
+	if err != nil {
+		t.Fatalf("failed to create GET /%s request: %v", key, err)
 	}
 
-	if tr.ComposeStack != nil {
-		err := tr.ComposeStack.Down(ctx, compose.RemoveOrphans(true), compose.RemoveImagesLocal)
-		if err != nil {
-			t.Logf("failed to tear down compose stack: %v", err)
-		}
+	client := &http.Client{
+		CheckRedirect: func(req *http.Request, via []*http.Request) error {
+			// stop following redirects
+			return http.ErrUseLastResponse
+		},
 	}
+
+	resp, err := client.Do(req)
+	if err != nil {
+		t.Fatalf("failed to make GET /%s request: %v", key, err)
+	}
+
+	return resp
 }
 
-func getKeyDbUrl(ctx context.Context, composeStack *compose.DockerCompose) (string, error) {
-	keyDbContainer, err := composeStack.ServiceContainer(ctx, "postgres")
+func parseJSONResponse[T any](t *testing.T, resp *http.Response) *T {
+	t.Helper()
+
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		return "", fmt.Errorf("failed to get key db container: %w", err)
+		t.Fatalf("failed to read response body: %v", err)
 	}
 
-	keyDbPort, err := keyDbContainer.MappedPort(ctx, "5432")
+	var result T
+	err = json.Unmarshal(body, &result)
 	if err != nil {
-		return "", fmt.Errorf("failed to get key db mapped port: %w", err)
+		t.Fatalf("failed to parse JSON response: %v. Body: %s", err, string(body))
 	}
 
-	keyDbHost, err := keyDbContainer.Host(ctx)
-	if err != nil {
-		return "", fmt.Errorf("failed to get key db host: %w", err)
-	}
-
-	return fmt.Sprintf(
-		"postgres://testuser:password@%s:%s/keydb?sslmode=disable",
-		keyDbHost,
-		keyDbPort.Port(),
-	), nil
-}
-
-func getAppUrl(ctx context.Context, composeStack *compose.DockerCompose) (string, error) {
-	appContainer, err := composeStack.ServiceContainer(ctx, "url-shortener-server")
-	if err != nil {
-		return "", fmt.Errorf("failed to get app container: %w", err)
-	}
-
-	appPort, err := appContainer.MappedPort(ctx, "3001")
-	if err != nil {
-		return "", fmt.Errorf("failed to get app mapped port : %w", err)
-	}
-
-	appHost, err := appContainer.Host(ctx)
-	if err != nil {
-		return "", fmt.Errorf("failed to get app host: %w", err)
-	}
-
-	return fmt.Sprintf("http://%s:%s", appHost, appPort.Port()), nil
+	return &result
 }
